+++ modules/util/portal-tools-sample-sql-builder/src/main/resources/com/liferay/portal/tools/sample/sql/builder/dependencies/groups.ftl
@@ -34,7 +34,7 @@
 	<#include "wiki.ftl">
 
 	<@insertDLFolder
-		_ddmStructureId=dataFactory.defaultDLDDMStructureId
+		_ddmStructureId=dlDataFactory.defaultDLDDMStructureId
 		_dlFolderDepth=1
 		_groupId=groupId
 		_parentDLFolderId=0
